Updated a bit of the document, more changes soon.
I have attempted to mitigate some of the changes that have occurred automatically. I will create the table of content above and have placed a template of how it should be structured in order to me to follow in case I get lost in the future. 

I will likely change all the '>' later on as in my head it made sense but when reading it makes it just a headache to navigate the document.